Move CacheEvents and vairants into their own file
